Create status.js module
